extracted connection from seed-test.ts

+++ backend/db/seeds/seed-test.ts
@@ -1,25 +1,14 @@
-//note- production will need a seed with collections but no data
-import { MongoClient } from "mongodb";
-import dotenv from "dotenv";
+import {db, client} from "../connection";
 import userData from "../data/test/users";
 
-// Connection URL
-let client:any;
-const ENV = process.env.NODE_ENV || "test";
-console.log(ENV)
-dotenv.config();
-if (ENV === "test") {
-client = new MongoClient(`${process.env.MONGODB_URL}`);
-};
-
-
-//schemas
-import mongoose from "mongoose";
-import {Schema} from "mongoose";
-import test from "node:test";
-var bcrypt = require('bcrypt-nodejs');
-
 /*
+//schemas
+//import mongoose from "mongoose";
+import { Schema } from "mongoose";
+import test from "node:test";
+var bcrypt = require("bcrypt-nodejs");
+
+
 var userSchema = new Schema({
   username: String,
   password: String,
@@ -47,46 +36,34 @@ const isValid = bcrypt.compareSync("password", testUser.password);
 console.log("check valid password", isValid)
 */
 
-
-
-// Database Name
-const dbName = "char-portfolio-test";
-export const db = client.db(dbName);
-
-
-
-async function seed() {
-  
+export async function seed() {
   // Use connect method to connect to the server
   await client.connect();
   console.log("Connected successfully to server");
- 
 
   //drop collections
-  await  db.dropCollection("users")
- 
+  let list = await db.listCollections().toArray();
+  await list.forEach(async (data: any) => {
+    if (data.name === "users") {
+      console.log("dropping");
+      await db.dropCollection("users");
+    }
+  });
+
   //create collections
   const users = db.collection("users");
-
 
   //insert users data
   const insertResult = await users.insertMany(userData);
   //console.log('Inserted documents =>', insertResult);
 
   const filteredDocs = await users.find({}).toArray();
-//console.log('Found documents  =>', filteredDocs);
-
-
-  return "done.";
-  
-}
+  //console.log('Found documents  =>', filteredDocs);
+ // await client.close()
+  //return "done.";
+};
 
 seed()
-  .then(console.log)
-  .catch(console.error)
- // .finally(() => client.close());
-
-  export default seed;
-
-
-
+  //.then(console.log)
+  //.catch(console.error)
+  //.finally(() => client.close());
